Rename media.info.response.ts to media.repository.info.response.ts
