+++ src/responses/index.ts
@@ -27,5 +27,4 @@ export * from './upload.repository.photo.response';
 export * from './user.feed.response';
 export * from './user.repository.info.response';
 export * from './user.repository.search.response';
-export * from './media.info.response';
-
+export * from './media.repository.info.response';
